requirements.txt
streamlit
ultralytics
numpy
opencv-python
pillow
